requirements.txt
python-binance
python-dotenv
numpy
TA-Lib
websocket
websocket-client
matplotlib
discord